support JSON3.read and JSON3.write, and declare structtypes
